refactor: split PlayerProfileView and add AGENTS.md for Codex
